// src/modules/esop/translate.enum.ts
export enum ESOPKeys {
    NOT_FOUND = "esop.NOT_FOUND",
    CREATION_ERROR = "esop.CREATION_ERROR",
    CREATED_SUCCESSFULLY = "esop.CREATED_SUCCESSFULLY",
    UPDATE_ERROR = "esop.UPDATE_ERROR",
    UPDATED_SUCCESSFULLY = "esop.UPDATED_SUCCESSFULLY",
    TOTAL_SHARES_UPDATED_SUCCESSFULLY = "esop.TOTAL_SHARES_UPDATED_SUCCESSFULLY",
    CANT_DELETE = "esop.CANT_DELETE",
    TOTAL_SHARES_ERROR = "esop.TOTAL_SHARES_ERROR",
    DELETED_SUCCESSFULLY = "esop.DELETED_SUCCESSFULLY",
    COMPANY_ACCESS_DENIED = "esop.COMPANY_ACCESS_DENIED",
    WRONG_LIST = "esop.WRONG_LIST",
    SHARES_PERCENTAGE_ERROR = "esop.SHARES_PERCENTAGE_ERROR",
    INSUFFICIENT_TOTAL_SHARES_ERROR = "esop.INSUFFICIENT_TOTAL_SHARES_ERROR",
}